# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangeTenantStatusDto": {
        "properties": {
          "reason": {
            "maxLength": 1000,
            "type": "string"
          },
          "status": {
            "enum": ["DRAFT", "ACTIVE", "SUSPENDED", "CLOSED"],
            "type": "string"
          }
        },
        "required": ["status"],
        "type": "object"
      },
      "CorrectFirstAdminPhoneDto": {
        "properties": {
          "firstAdminPhone": {
            "example": "13900139000",
            "type": "string"
          }
        },
        "required": ["firstAdminPhone"],
        "type": "object"
      },
...
      "CreatePlatformTenantDto": {
        "properties": {
          "code": {
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "type": "string"
          },
          "firstAdminPhone": {
            "example": "13800138000",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": ["name", "code", "firstAdminPhone"],
        "type": "object"
      },
...
      "FirstAdminInvitationResponseDto": {
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "enum": ["TENANT_ADMIN"],
            "type": "string"
          },
          "status": {
            "enum": ["PENDING", "ACCEPTED", "DECLINED", "REVOKED", "EXPIRED"],
            "type": "string"
          },
          "targetPhone": {
            "example": "+8613800138000",
            "type": "string"
          }
        },
        "required": ["id", "targetPhone", "role", "status", "expiresAt"],
        "type": "object"
      },
...
      "PlatformTenantPageResponseDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlatformTenantResponseDto"
            },
            "type": "array"
          },
          "limit": {
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "page": {
            "minimum": 1,
            "type": "number"
          },
          "total": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": ["items", "page", "limit", "total"],
        "type": "object"
      },
      "PlatformTenantResponseDto": {
        "properties": {
          "activatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "activeAdminCount": {
            "minimum": 0,
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "firstAdminInvitation": {
            "$ref": "#/components/schemas/FirstAdminInvitationResponseDto"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "enum": ["DRAFT", "ACTIVE", "SUSPENDED", "CLOSED"],
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "name", "code", "status", "activeAdminCount"],
        "type": "object"
      },
      "PlatformTenantSummaryDto": {
        "properties": {
          "active": {
            "minimum": 0,
            "type": "number"
          },
          "closed": {
            "minimum": 0,
            "type": "number"
          },
          "draft": {
            "minimum": 0,
            "type": "number"
          },
          "suspended": {
            "minimum": 0,
            "type": "number"
          },
          "total": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": ["total", "draft", "active", "suspended", "closed"],
        "type": "object"
      },
...
      "TenantNameConflictResponseDto": {
        "properties": {
          "count": {
            "minimum": 0,
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["name", "count"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "crm_session": {
        "in": "cookie",
        "name": "crm_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "多租户 CRM 平台 REST API",
    "title": "Multi-tenant CRM API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/platform/tenants": {
      "get": {
        "operationId": "TenantsController_list",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": ["DRAFT", "ACTIVE", "SUSPENDED", "CLOSED"],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformTenantPageResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["platform-tenants"]
      },
      "post": {
        "operationId": "TenantsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlatformTenantDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformTenantResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["platform-tenants"]
      }
    },
    "/api/v1/platform/tenants/name-conflicts": {
      "get": {
        "operationId": "TenantsController_nameConflicts",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantNameConflictResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["platform-tenants"]
      }
    },
    "/api/v1/platform/tenants/summary": {
      "get": {
        "operationId": "TenantsController_summarize",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformTenantSummaryDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["platform-tenants"]
      }
    },
    "/api/v1/platform/tenants/{tenantId}": {
      "get": {
        "operationId": "TenantsController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformTenantResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["platform-tenants"]
      }
    },
...
    "/api/v1/platform/tenants/{tenantId}/first-admin-invitation/renew": {
      "post": {
        "operationId": "TenantsController_renewFirstAdminInvitation",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformTenantResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["platform-tenants"]
      }
    },
    "/api/v1/platform/tenants/{tenantId}/first-admin-phone": {
      "patch": {
        "operationId": "TenantsController_correctFirstAdminPhone",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectFirstAdminPhoneDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformTenantResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["platform-tenants"]
      }
    },
    "/api/v1/platform/tenants/{tenantId}/status": {
      "patch": {
        "operationId": "TenantsController_changeStatus",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeTenantStatusDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformTenantResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "crm_session": []
          }
        ],
        "tags": ["platform-tenants"]
      }
    },
...
  },
  "servers": [],
...
}